Renamespace applications
