+++ config/services_test.yaml
@@ -35,16 +35,16 @@ services:
             $adminCreateUserUrl: '%route-admin-user-create%'
             $adminRevokeRefreshTokenUrl: '%route-admin-frontend-refresh-token-revoke%'
 
-    App\Tests\Services\AbstractBaseApplication:
+    App\Tests\Services\Application\AbstractBaseApplication:
         abstract: true
 
-    App\Tests\Services\Application:
+    App\Tests\Services\Application\FunctionalApplication:
         public: true
-        parent: App\Tests\Services\AbstractBaseApplication
+        parent: App\Tests\Services\Application\AbstractBaseApplication
 
-    App\Tests\Services\IntegrationApplication:
+    App\Tests\Services\Application\IntegrationApplication:
         public: true
-        parent: App\Tests\Services\AbstractBaseApplication
+        parent: App\Tests\Services\Application\AbstractBaseApplication
 
     App\Tests\Services\RefreshTokenManager:
         public: true
