+++ tests/Functional/AbstractBaseWebTestCase.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace App\Tests\Functional;
 
-use App\Tests\Services\Application;
-use App\Tests\Services\ApplicationInterface;
+use App\Tests\Services\Application\ApplicationInterface;
+use App\Tests\Services\Application\FunctionalApplication;
 
 abstract class AbstractBaseWebTestCase extends AbstractBaseFunctionalTest
 {
@@ -17,8 +17,8 @@ abstract class AbstractBaseWebTestCase extends AbstractBaseFunctionalTest
 
         $client = static::createClient();
 
-        $application = self::getContainer()->get(Application::class);
-        \assert($application instanceof Application);
+        $application = self::getContainer()->get(FunctionalApplication::class);
+        \assert($application instanceof FunctionalApplication);
         $application->setClient($client);
 
         $this->application = $application;
