+++ tests/Integration/AbstractIntegrationTest.php
@@ -7,9 +7,9 @@ namespace App\Tests\Integration;
 use App\Entity\User;
 use App\Repository\UserRepository;
 use App\Services\ApiKeyFactory;
-use App\Tests\Services\ApplicationInterface;
+use App\Tests\Services\Application\ApplicationInterface;
+use App\Tests\Services\Application\IntegrationApplication;
 use App\Tests\Services\ApplicationResponseAsserter;
-use App\Tests\Services\IntegrationApplication;
 use App\Tests\Services\UserRemover;
 use Doctrine\ORM\EntityManagerInterface;
 use Gesdinet\JWTRefreshTokenBundle\Entity\RefreshToken;
