+++ tests/Services/Application/IntegrationApplication.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Services;
+namespace App\Tests\Services\Application;
 
+use App\Tests\Services\ApplicationRoutes;
 use GuzzleHttp\Psr7\Utils;
 use Psr\Http\Client\ClientInterface;
 use Psr\Http\Message\RequestFactoryInterface;
